Add files via upload
Here, I upload a Project of Employee Management System. An Employee Management System (EMS) is a software application or platform designed to streamline and automate various aspects of workforce management within an organization. The primary goal of an EMS is to enhance efficiency, accuracy, and productivity in managing employee-related tasks and information.